Version 0.2.9 - Added default.config capability
This version attempts to load default.config and only prompts
for input if it is unable to open the file.  You can also specify
'-s <file to open>' to load a different set of parameters or '-s'
to skip straight to the prompt without trying to open the defaults.

Other changes:
- all instance parameters are stored in the <inst> struct
- minor fixes

Note: Loading parameters from a file does not verify them.  Thus,
loading a bad default.config can cause the program to hang if
one of the specified parameters is not valid.