Version 0.3.3 - GPS Parser and Local Testing
This version corrects almost all bad functionality with the GPS
parser (see note below) and adds the local testing variant.  By
uncommenting #define _LOCALTEST_ and compiling against rsfile.cpp
instead of rs232.c, you're able to use the entire program without
a comport without modifying any code.  The rsfile library substitues
the necessary rs232 functions for local file operations: one for
data in, one for data out.  The program prompts for these files
when comport is opened (usually at the beginning of the program).

Known Issue:
If the datafile already contains data when the program begins,
the program with segfault when trying to write the HTML version.
This will be corrected with the next version.

Workaround - delete the datafile before each new start.